+++ tests/tests/main.cpp
@@ -28,7 +28,7 @@
 extern uint32_t GRAPHENE_TESTING_GENESIS_TIMESTAMP;
 
 boost::unit_test::test_suite* init_unit_test_suite(int argc, char* argv[]) {
-   std::srand(time(NULL));
+   std::srand(static_cast<unsigned int>(time(NULL)));
    std::cout << "Random number generator seeded to " << time(NULL) << std::endl;
    const char* genesis_timestamp_str = getenv("GRAPHENE_TESTING_GENESIS_TIMESTAMP");
    if( genesis_timestamp_str != nullptr )
